Update base command code
Before this commit, there was a `jobid` system, where each command would be downloaded and then processed, written to file once more, and then uploaded.

New things include that the file is no longer downloaded to a file, and is instead only kept in memory, and the same applies to the modified image. A `data_manipulator` field was added, allowing another function to be appended after `fn` in order to manipulate the command's data, allowing the user to change things such as the options of a command.

Also added a `Time took` field into the embed, and simplified `greyscale`